+++ sdk/ai/Azure.AI.Agents.Persistent/src/Custom/Streaming/StreamingUpdateCollection.cs
@@ -204,7 +204,12 @@ internal class StreamingUpdateCollection : CollectionResult<StreamingUpdate>
                         return false;
                     }
 
-                    var updates = StreamingUpdate.FromEvent(_events.Current);
+                    IEnumerable<StreamingUpdate> updates = StreamingUpdate.FromEvent(_events.Current);
+                    if (updates is null)
+                    {
+                        StreamingUpdateReason updateKind = StreamingUpdateReasonExtensions.FromSseEventLabel(_events.Current.EventType);
+                        throw new InvalidOperationException($"Unknown streaming update reason {updateKind}");
+                    }
                     _updates = updates.GetEnumerator();
 
                     if (_updates.MoveNext())